[feature/api-auth] : Add table column, refactoring, add token timelife
The final token have to be refreshed. The user can modify the option in ACP to make it infinite and it's pushed in api_keys table when the authentication is verified.
I did a little refactoring on the test files and the method authenticate which was highlighted with a TODO.

PHPBB3-11595